requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
feedparser==6.0.10
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
loguru==0.7.2
lunardate==0.2.0
MarkupSafe==2.1.3
nltk==3.8.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
snownlp==0.12.3
soupsieve==2.5
textblob==0.17.1
tqdm==4.66.1
urllib3==2.0.7
Werkzeug==3.0.1
